Calolator-Backend/requirements.txt
fastapi[standard]
transformers
torch
accelerate
peft
bitsandbytes
python-dotenv
scipy
pyngrok